+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## [Unreleased]
 
+### Enhancements
+
+* Integrate `libcurl` for HTTP requests, replacing external `curl`/`wget` tools
+
 ## v25.11.29
 
 ### Enhancements
@@ -16,7 +20,6 @@
 
 * Use `v2/submissions/` API endpoint to fetch and submit programs
 * Handle program submissions with `delete` mode
-* Integrate libcurl for HTTP requests, replacing external curl/wget tools
 
 ## v25.11.9